add new logs + update interaction_summary
