+++ src/interaction/include/exaDEM/interaction/classifier.hpp
@@ -1,11 +1,25 @@
-
 #pragma once
 
 #include <exaDEM/interaction/interaction.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
+#include <exaDEM/itools/buffer.hpp>
 
 namespace exaDEM
 {
+
+  template<typename GridT>
+    inline bool filter_duplicates(const GridT& G, const exaDEM::Interaction& I)
+    {
+      if(I.type < 4) // polyhedron - polyhedron or sphere - sphere
+      {
+        if(G.is_ghost_cell(I.cell_j) && I.id_i > I.id_j)
+        {
+          return false;
+        }
+      }
+      return true;
+    }
+
 
   struct InteractionWrapper
   {
@@ -25,27 +39,6 @@ namespace exaDEM
   };
 
 
-  struct analysis_buffers
-  {
-    template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
-    VectorT<double> dn;
-    VectorT<Vec3d>  cp;
-    VectorT<Vec3d>  fn;
-    VectorT<Vec3d>  ft;
-
-    void resize (const int size)
-    {
-      if( size != 0 ) 
-      {
-
-        dn.resize(size);
-        cp.resize(size);
-        fn.resize(size);
-        ft.resize(size);
-      }
-    }
-  };
-
   /**
    * @brief Classifier for managing interactions categorized into different types.
    *
@@ -56,8 +49,8 @@ namespace exaDEM
   {
     static constexpr int types = 13;
     template <typename T> using VectorT =  onika::memory::CudaMMVector<T>;
-    std::vector<VectorT<exaDEM::Interaction>> waves; ///< Storage for interactions categorized by type.
-    std::vector<analysis_buffers> buffers; ///< Storage for analysis. Empty if there is no analysis
+    std::vector< VectorT< exaDEM::Interaction >> waves; ///< Storage for interactions categorized by type.
+    std::vector< itools::interaction_buffers > buffers; ///< Storage for analysis. Empty if there is no analysis
 
     /**
      * @brief Default constructor.
@@ -89,6 +82,7 @@ namespace exaDEM
      * @return Reference to the CUDA memory-managed vector storing interactions of the specified type.
      */
     VectorT<exaDEM::Interaction>& get_wave(size_t id) {return waves[id];}
+    const VectorT<exaDEM::Interaction>& get_wave(size_t id) const {return waves[id];}
 
     /**
      * @brief Retrieves the pointer and size of the data stored in the CUDA memory-managed vector for a specific type.
@@ -103,176 +97,120 @@ namespace exaDEM
       return {data_ptr, data_size};
     }
 
-    /**
-     * @brief This function returns the total number of classifier interactions.
-     */
-    uint64_t number_of_interactions(bool sym)
+    const std::pair< const exaDEM::Interaction* const , const size_t> get_info(size_t id) const
     {
-      uint64_t res = 0 ;
-#pragma omp parallel for reduction(+: res )
-      for(size_t i = 0 ; i < waves.size() ; i++ )
-      {
-        auto [ptr, size] = get_info(i);
-        if( i < 4 && sym == true ) res += 2 * size;
-        else res += size;
-      }
-      ldbg << " number_of_interactions " << res << std::endl;
-      return res;
+      const unsigned int  data_size = onika::cuda::vector_size( waves[id]);
+      const exaDEM::Interaction* const data_ptr = onika::cuda::vector_data( waves[id] );
+      return {data_ptr, data_size};
     }
 
-    std::tuple<double*, Vec3d*,Vec3d*,Vec3d*> buffer_p(int id)
-    {
-      auto& analysis = buffers[id]; 
-      // fit size if needed
-      const size_t size = onika::cuda::vector_size( waves[id] );
-      analysis.resize(size);
-      double* const dnp = onika::cuda::vector_data( analysis.dn ); 
-      Vec3d*  const cpp = onika::cuda::vector_data( analysis.cp ); 
-      Vec3d*  const fnp = onika::cuda::vector_data( analysis.fn ); 
-      Vec3d*  const ftp = onika::cuda::vector_data( analysis.ft );
-      return {dnp, cpp, fnp, ftp}; 
-    }
+		std::tuple<double*, Vec3d*,Vec3d*,Vec3d*> buffer_p(int id)
+		{
+			auto& analysis = buffers[id]; 
+			// fit size if needed
+			const size_t size = onika::cuda::vector_size( waves[id] );
+			analysis.resize(size);
+			double* const dnp = onika::cuda::vector_data( analysis.dn ); 
+			Vec3d*  const cpp = onika::cuda::vector_data( analysis.cp ); 
+			Vec3d*  const fnp = onika::cuda::vector_data( analysis.fn ); 
+			Vec3d*  const ftp = onika::cuda::vector_data( analysis.ft );
+			return {dnp, cpp, fnp, ftp}; 
+		}
 
+		/**
+		 * @brief Returns the number of interaction types managed by the classifier.
+		 *
+		 * @return Number of interaction types.
+		 */
+		size_t number_of_waves() { assert(types == waves.size()) ; return types;}
+		const size_t number_of_waves() const { assert(types == waves.size()) ; return types;}
 
-    /**
-     * @brief This functions returns the number of active interactions; i.e. if two particles are overlapped by themselves.
-     */
-    uint64_t get_active_interactions(bool sym)
-    {
-      // Warning, this function has to be called when the analysis buffer is set (in hooke operators).
-      // TODO : Implement a GPU version
-      uint64_t res = 0;
-      for( size_t i = 0 ; i < types ; i++ )
-      {
-        // number of iteractions per wave
-        auto& analysis = buffers[i]; 
-        const double* const dnp = onika::cuda::vector_data( analysis.dn ); 
-        const size_t size = onika::cuda::vector_size( analysis.dn );
-        uint64_t coef = 1;
-        if( i < 4 && sym == true) coef = 2; 
-#pragma omp parallel for reduction(+: res)
-        for( size_t j = 0 ; j < size ; j++ )
-        {
-          if( dnp[j] < 0 ) res += coef;
-        }
-      }
-      ldbg << " get_active_interactions " << res << std::endl;
-      return res;
-    }
-
-    double get_min_dn ()
-    {
-      // TODO : Implement a GPU version
-      double res = 0;
-      for( size_t i = 0 ; i < types ; i++ )
-      {
-        // number of iteractions per wave
-        auto& analysis = buffers[i]; 
-        const double* const dnp = onika::cuda::vector_data( analysis.dn ); 
-        const size_t size = onika::cuda::vector_size( analysis.dn );
-#pragma omp parallel for reduction(min: res)
-        for( size_t j = 0 ; j < size ; j++ )
-        {
-          const double dn = dnp[j];
-          if( dn < res ) res = dn;
-        }
-      }
-      return res;
-    }
-
-    /**
-     * @brief Returns the number of interaction types managed by the classifier.
-     *
-     * @return Number of interaction types.
-     */
-    size_t number_of_waves() { assert(types == waves.size()) ; return waves.size();}
-
-    /**
-     * @brief Classifies interactions into categorized waves based on their types.
-     *
-     * This function categorizes interactions into different waves based on their types,
-     * utilizing the `waves` vector in the `Classifier` struct. It resets existing waves,
-     * calculates the number of interactions per wave, and then stores interactions
-     * accordingly.
-     *
-     * @param ges Reference to the GridCellParticleInteraction object containing interactions to classify.
-     */
-    void classify(GridCellParticleInteraction& ges, size_t* idxs, size_t size)
-    {
-      reset_waves(); // Clear existing waves
-      auto& ces = ges.m_data; // Reference to cells containing interactions
+		/**
+		 * @brief Classifies interactions into categorized waves based on their types.
+		 *
+		 * This function categorizes interactions into different waves based on their types,
+		 * utilizing the `waves` vector in the `Classifier` struct. It resets existing waves,
+		 * calculates the number of interactions per wave, and then stores interactions
+		 * accordingly.
+		 *
+		 * @param ges Reference to the GridCellParticleInteraction object containing interactions to classify.
+		 */
+		void classify(GridCellParticleInteraction& ges, size_t* idxs, size_t size)
+		{
+			reset_waves(); // Clear existing waves
+			auto& ces = ges.m_data; // Reference to cells containing interactions
 
 #pragma omp parallel
-      {
-        std::array<std::vector<exaDEM::Interaction>,types> tmp; ///< Storage for interactions categorized by type.
+			{
+				std::array<std::vector<exaDEM::Interaction>,types> tmp; ///< Storage for interactions categorized by type.
 #pragma omp for schedule(dynamic) nowait
-        for(size_t c = 0 ; c < size ; c++)
-        {
-          auto& interactions = ces[idxs[c]];
-          const unsigned int  n_interactions_in_cell = interactions.m_data.size();
-          exaDEM::Interaction* const __restrict__ data_ptr = onika::cuda::vector_data( interactions.m_data );
-          // Place interactions into their respective waves  
-          for( size_t it = 0; it < n_interactions_in_cell ; it++ )
-          {
-            Interaction& item = data_ptr[it];
-            const int t = item.type;
-            tmp[t].push_back(item);
-          }
-        }
+				for(size_t c = 0 ; c < size ; c++)
+				{
+					auto& interactions = ces[idxs[c]];
+					const unsigned int  n_interactions_in_cell = interactions.m_data.size();
+					exaDEM::Interaction* const __restrict__ data_ptr = onika::cuda::vector_data( interactions.m_data );
+					// Place interactions into their respective waves  
+					for( size_t it = 0; it < n_interactions_in_cell ; it++ )
+					{
+						Interaction& item = data_ptr[it];
+						const int t = item.type;
+						tmp[t].push_back(item);
+					}
+				}
 
-        for(int w = 0 ; w < types ; w++ )
-        {
+				for(int w = 0 ; w < types ; w++ )
+				{
 #pragma omp critical
-          {
-            waves[w].insert(waves[w].end(), tmp[w].begin(), tmp[w].end());
-          }
-        }
-      }
-    }
+					{
+						waves[w].insert(waves[w].end(), tmp[w].begin(), tmp[w].end());
+					}
+				}
+			}
+		}
 
-    /**
-     * @brief Restores friction and moment data for interactions from categorized waves to cell interactions.
-     *
-     * This function restores friction and moment data from categorized waves back to their corresponding
-     * interactions in cell data (`ges.m_data`). It iterates through each wave, retrieves interactions
-     * with non-zero friction and moment from the wave, and updates the corresponding interaction in the
-     * cell data.
-     *
-     * @param ges Reference to the GridCellParticleInteraction object containing cell interactions.
-     */
-    void unclassify(GridCellParticleInteraction& ges)
-    {
-      Vec3d null = {0,0,0};
-      auto& ces = ges.m_data; // Reference to cells containing interactions
-                              // Iterate through each wave
-      for(int w = 0 ; w < types ; w++)
-      {
-        auto& wave = waves[w];
-        const unsigned int n1 = wave.size();
-        // Parallel loop to process interactions within a wave
+		/**
+		 * @brief Restores friction and moment data for interactions from categorized waves to cell interactions.
+		 *
+		 * This function restores friction and moment data from categorized waves back to their corresponding
+		 * interactions in cell data (`ges.m_data`). It iterates through each wave, retrieves interactions
+		 * with non-zero friction and moment from the wave, and updates the corresponding interaction in the
+		 * cell data.
+		 *
+		 * @param ges Reference to the GridCellParticleInteraction object containing cell interactions.
+		 */
+		void unclassify(GridCellParticleInteraction& ges)
+		{
+			Vec3d null = {0,0,0};
+			auto& ces = ges.m_data; // Reference to cells containing interactions
+															// Iterate through each wave
+			for(int w = 0 ; w < types ; w++)
+			{
+				auto& wave = waves[w];
+				const unsigned int n1 = wave.size();
+				// Parallel loop to process interactions within a wave
 #pragma omp parallel for
-        for(size_t it = 0 ; it < n1 ; it++) 
-        {
-          exaDEM::Interaction& item1 = wave[it];
-          // Check if interaction in wave has non-zero friction and moment
-          if( item1.friction != null || item1.moment != null)
-          { 
-            auto& cell = ces[item1.cell_i];
-            const unsigned int  n2 = onika::cuda::vector_size( cell.m_data );
-            exaDEM::Interaction* data_ptr = onika::cuda::vector_data( cell.m_data );
-            // Iterate through interactions in cell to find matching interaction
-            for(size_t it2 = 0; it2 < n2 ; it2++)
-            {
-              exaDEM::Interaction& item2 = data_ptr[it2];
-              if(item1 == item2)
-              {
-                item2.update_friction_and_moment(item1);
-                break;
-              }
-            }
-          }
-        }
-      }
-    }
-  };
+				for(size_t it = 0 ; it < n1 ; it++) 
+				{
+					exaDEM::Interaction& item1 = wave[it];
+					// Check if interaction in wave has non-zero friction and moment
+					if( item1.friction != null || item1.moment != null)
+					{ 
+						auto& cell = ces[item1.cell_i];
+						const unsigned int  n2 = onika::cuda::vector_size( cell.m_data );
+						exaDEM::Interaction* data_ptr = onika::cuda::vector_data( cell.m_data );
+						// Iterate through interactions in cell to find matching interaction
+						for(size_t it2 = 0; it2 < n2 ; it2++)
+						{
+							exaDEM::Interaction& item2 = data_ptr[it2];
+							if(item1 == item2)
+							{
+								item2.update_friction_and_moment(item1);
+								break;
+							}
+						}
+					}
+				}
+			}
+		}
+	};
 }
